+++ app/[shape]/page.tsx
@@ -154,10 +154,10 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
       title: title,
       description: description,
       type: 'website',
-      url: `https://caratcompare.co/${shape}`,
+      url: `https://www.caratcompare.co/${shape}`,
     },
     alternates: {
-      canonical: `https://caratcompare.co/${shape}`,
+      canonical: `https://www.caratcompare.co/${shape}`,
     },
   };
 }